Enhance clang-exception-scan with comprehensive analysis statistics
- Replaced the function definition count reporting with a new combined analysis statistics report, providing detailed metrics on non-system-header function definitions, try blocks, catch handlers, throw expressions, and calls within try blocks.
- Updated GlobalExceptionInfo to include additional atomic counters for tracking try blocks, catch handlers, throw expressions, and calls potentially within try blocks, ensuring accurate and thread-safe metrics.
- Refactored reporting functions in CollectExceptionInfo to utilize the new analysis statistics, improving clarity and consistency in output.
- Enhanced unit tests to validate the correctness of the new statistics reporting, ensuring comprehensive coverage of exception handling constructs.

+++ clang/unittests/Tooling/CallGraphGeneratorTest.cpp
@@ -542,7 +542,7 @@ TEST_F(CallGraphGeneratorTest, CollectOperatorNewDelete) {
   const std::string Code = R"(
     class MyClass {
     public:
-      void* operator new(size_t size) { return ::operator new(size); }
+      void* operator new(unsigned long size) { return ::operator new(size); }
       void operator delete(void* ptr) { ::operator delete(ptr); }
     };
     void func() {
